Begin implemenation of integrals Allocation of rate arrays
