+++ src/Dipole.h
@@ -10,7 +10,7 @@ namespace Dipole
 	// occ - orbitals occupancy
 	double DsigmaBEB(double T, double W, double B, double t, int occ);
 	double sigmaBEB(double T, double B, double u, int occ);
-	// Int_0^(p*p/2 - B) dW W d(sigmaBED)/dW - first moment of a secondary electron energy.
+	// Int_0^(T - B) dW W d(sigmaBED)/dW - total energy absorbance
 	double sigmaBEBw1(double T, double B, double u, int occ);
 };
