+++ src/ElectronSolver.cpp
@@ -5,6 +5,7 @@
 #include <sys/stat.h>
 #include <algorithm>
 
+
 void PhotonFlux::set_parameters(double fluence, double fwhm){
     // The photon flux model
     // Gaussian model A e^-t^2/B
@@ -13,7 +14,7 @@ void PhotonFlux::set_parameters(double fluence, double fwhm){
 }
 
 inline double PhotonFlux::operator()(double t){
-    // Returns flux at time t
+    // Returns flux at time t (same units as fluence)
     return A*exp(-t*t/B);
 }
 
@@ -22,43 +23,19 @@ void PhotonFlux::save(const vector<double>& Tvec, const std::string& fname){
     cout << "[ Flux ] Saving to file "<<fname<<"..."<<endl;
     f.open(fname);
     f << "# Free electron dynamics"<<endl;
-    f << "# Time (fs) | Intensity [ UNITS ] " <<endl;
+    f << "# Time (fs) | Intensity (J/cm2)" <<endl;
     for (auto& t : Tvec){
         f << t << " "<< this->operator()(t)<<std::endl;
     }
     f.close();
 }
 
-Weight::Weight(size_t _size){
-    size = _size;
-    W = (double *) malloc(sizeof(double)*size*size); // allocate the memory
-    std::fill(W, W+size*size, 0);
-}
-
-Weight::~Weight(){
-    free(W);
-}
-
-double Weight::from(size_t idx){
-    double x=0;
-    for (size_t i = 0; i < size; i++) {
-        x += operator()(i, idx);
-    }
-    return x;
-}
-
-double Weight::to(size_t idx){
-    double x=0;
-    for (size_t i = 0; i < size; i++) {
-        x += operator()(idx, i);
-    }
-    return x;
-}
 
 ElectronSolver::ElectronSolver(const char* filename, ofstream& log) :
-    MolInp(filename, log), pf(width, 100000*fluence)
+    MolInp(filename, log), pf(width*Constant::fs_in_au, 100000*fluence*Constant::Fluence_in_au)
 {
-    timespan = this->width*10;
+    timespan = this->width*10*Constant::fs_in_au;
+    precompute_gamma_coeffs();
 }
 
 state_type ElectronSolver::get_ground_state() {
@@ -81,7 +58,7 @@ void ElectronSolver::compute_cross_sections(std::ofstream& _log, bool recalc) {
     // Set up the rate equations (setup called from parent Adams_BM)
     this->setup(get_ground_state(), this->timespan/num_time_steps);
     cout<<"Using timestep "<<this->dt<<" fs"<<endl;
-};
+}
 
 void ElectronSolver::solve(){
     if (!hasRates) {
@@ -95,16 +72,43 @@ Use ElectronSolver::compute_cross_sections(log)\n" << endl;
     this->iterate(-timespan/2, this->num_time_steps); // Inherited from ABM
 }
 
+void ElectronSolver::precompute_gamma_coeffs(){
+    size_t N = Distribution::size;
+    size_t gamma_dim = Store[a].Photo.size();
+    for (size_t a = 0; a < s.atomP.size(); a++) {
+        for (size_t k=0; k<N*(N-1)/2; k++){
+            RATE_EII[a][k] = new Eigen::SparseMatrix<double>(gamma_dim, gamma_dim);
+        }
+        for (size_t n=0; n<N; n++){
+            RATE_EII[a][n] = new Eigen::SparseMatrix<double>(gamma_dim, gamma_dim);
+
+            for (auto& eii: Store[a].EIIparams){
+                Eigen::SparseMatrix<double> Gamma(gamma_dim, gamma_dim);
+                Gamma_eii(Gamma, eii, n, a);
+                *RATE_EII[a][n] += Gamma;
+                for (size_t m=0; m<n; m++){
+                    size_t k = (N*(N+1)/2) - (N-m)*(N-m-1)/2 + n - m - 1
+                    // k = N-1... N(N+1)/2
+                    Gamma_tbr(Gamma, eii, n, m, a);
+                    *RATE_TBR[a][k] += Gamma;
+                }
+                Gamma_tbr(Gamma, eii, n, n, a);
+                *RATE_TBR[a][n] += Gamma;
+            }
+        }
+    }
+}
 
 // The Big One: Incorporates all of the right hand side to the global
 // d/dt P[i] = \sum_i=1^N W_ij - W_ji P[j]
 // d/dt f = Q_B[f](t)
+// system
 void ElectronSolver::sys(const state_type& s, state_type& sdot, const double t){
     sdot=0;
     for (size_t a = 0; a < s.atomP.size(); a++) {
 
         // create an appropriately-sized W[i][j]
-        Weight W(s.atomP[a].size());
+        Eigen::MatrixXd W(s.atomP[a].size(), s.atomP[a].size());
 
         // PHOTOIONISATION
         double J = pf(t); // photon flux in uhhhhh [TODO: UNITS]
@@ -126,18 +130,19 @@ void ElectronSolver::sys(const state_type& s, state_type& sdot, const double t){
         }
 
         // EII / TBR
-        for ( auto& ep : Store[a].EIIparams) {
-            for (size_t i = 0; i < ep.fin.size(); i++) {
-                // Number of Electrons
-                // INTEGRAL 0 infty n_a/2 (2e/m) de/m  f(e) sigmaBEB(e)
-                // n_a num atoms
-                double n, dN;
-                n = Store[a].nAtoms;
-                dN  = n*(s.F.eii_int(ep, i) + n*s.F.tbr_int(ep, i));
-                W(ep.fin[i], i) += dN;
-
+        double N = Distribution::size;
+        for (size_t n=0; n<N; n++){
+            W += RATE_EII[a][n]*s.F.f[n];
+            // exploit the symmetry: goofy indexing to only store the upper triangular part.
+            for (size_t m=0; m<n; m++){
+                size_t k = (N*(N+1)/2) - (N-m)*(N-m-1)/2 + n - m - 1
+                // k = N-1... N(N+1)/2
+                W += RATE_TBR[a][k]*s.F.f[n]*s.F.f[m]*2;
             }
+            // the diagonal
+            W += RATE_TBR[a][n]*s.F.f[n]*s.F.f[n];
         }
+
 
 
         const state_type::bound_t& P = s.atomP[a];
@@ -153,10 +158,16 @@ void ElectronSolver::sys(const state_type& s, state_type& sdot, const double t){
             for (size_t j = i+1; j < P.size(); j++) {
                 Pdot[i] += W(i, j) * P[j] - W(j, i) * P[i];
             }
+
         }
+
+        // Changes in F
+        sdot.F.add_Qeii(a, s.F, P);
+        sdot.F.add_Qtbr(a, s.F, P);
+
     }
-    // Changes in F
     sdot.F.add_Qee(s.F); // Electron-electon repulsions
+
 }
 
 // IO functions
@@ -178,7 +189,7 @@ void ElectronSolver::saveFree(const std::string& fname){
     f.open(fname);
     f << "# Free electron dynamics"<<endl;
     f << "# Time (fs) | Density [ UNITS ] @ energy:" <<endl;
-    f << "#           | "<<Distribution::get_energies("eV")<<endl;
+    f << "#           | "<<Distribution::get_energies_eV()<<endl;
 
     assert(y.size() == t.size());
     for (int i=0; i<y.size(); i++){
@@ -214,3 +225,31 @@ void ElectronSolver::saveBound(const std::string& dir){
     }
 
 }
+
+/*
+Weight::Weight(size_t _size){
+    size = _size;
+    W = (double *) malloc(sizeof(double)*size*size); // allocate the memory
+    std::fill(W, W+size*size, 0);
+}
+
+Weight::~Weight(){
+    free(W);
+}
+
+double Weight::from(size_t idx){
+    double x=0;
+    for (size_t i = 0; i < size; i++) {
+        x += operator()(i, idx);
+    }
+    return x;
+}
+
+double Weight::to(size_t idx){
+    double x=0;
+    for (size_t i = 0; i < size; i++) {
+        x += operator()(idx, i);
+    }
+    return x;
+}
+*/
