+++ src/IntegrateRateEquation.cpp
@@ -289,84 +289,6 @@ IntegrateRateEquation::IntegrateRateEquation(vector<double> &dT, vector<double> 
 	}
 }
 
-/*
-// Rate equations for single chemical element, with non-thermal plasma.
-IntegrateRateEquation::IntegrateRateEquation(vector<double> &dT, vector<double> &T, RateData::Atom & Store, NTPlasma & Elecs, vector<double> InitCond, const vector<double>& Intensity) :
-    dt(dT), t(T), f(Intensity), store(Store)
-{
-	// // Compute time-evolution of the energy distribution
-	// // InitCond defines number of states and initial values for p.
-    //
-    // // Implementation of Morgan and Penetrante's ELENDIF Algorithm
-    // // Morgan and Penetrante - 1990 - ELENDIF A time-dependent Boltzmann
-    // // solver for partially ionized plasmas
-    //
-    // // Plasma Equations, recast in terms of number density distribution:
-    // // (e means energy, density is n)
-    // // Assume external field in z-direction,
-    // // f(v_z, v_rho) = f0(v_z, v_rho) + v_z*f1(v_z, v_rho)
-    //
-    // // n(e) = n_e f
-    // // dn(e)/dt = -J_f
-    // // Q[n] =
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    // //
-    //
-    //
-    //
-    // adams_n = 5;
-    // // Resize t, f arrays
-    // t.resize(dt.size());
-    //
-    // // Size of P_\lambda array, number of elemental configurations
-    // int tot_p_size = 0;
-	// int p_size = adams_n + 1;
-	// for (auto& elem: Store) tot_p_size += elem.num_conf;
-    //
-	// dpdt.resize(tot_p_size);
-	// p.resize(tot_p_size);
-    //
-	// for (int i = 0; i < tot_p_size; i++) {
-	// 	dpdt[i].resize(p_size, 0.);
-	// 	p[i].resize(p_size, 0.);
-	// }
-    //
-	// int init_p = 0;
-	// for (auto& elem: Store) {
-	// 	p[init_p][0] = 1.;
-	// 	init_p += elem.num_conf;
-	// }
-    //
-	// vector<vector<double*>> map_p(Store.size());
-	// vector<vector<double*>> map_dpdt(Store.size());
-	// init_p = 0;
-	// for (int a = 0; a < Store.size(); a++) {
-	// 	for (int i = 0; i < Store[a].num_conf; i++) {
-	// 		map_p[a].push_back(p[i + init_p].data());
-	// 		map_dpdt[a].push_back(dpdt[i + init_p].data());
-	// 	}
-	// 	init_p += Store[a].num_conf;
-	// }
-}
-
-*/
-
 // Rate equations for a molecule (thermal plasma).
 IntegrateRateEquation::IntegrateRateEquation(vector<double> &dT, vector<double> &T, vector<RateData::Atom> & Store, Plasma & Elecs, const vector<double>& Intensity) :
     dt(dT), t(T), f(Intensity), store(Store[0])
