+++ src/RateSystem.h
@@ -7,97 +7,16 @@
 #include <iostream>
 #include "Constant.h"
 #include "AdamsIntegrator.hpp"
+#include "FreeDistribution.h"
 
 
 
-/*
-class state_type :
-    boost::additive1< state_type ,
-    boost::additive2< state_type , double ,
-    boost::multiplicative2< state_type , double > > >
-    */
-
-// Represents a statistical distribution of electrons.
-// grid defines spacing in Hartree.
-class Distribution
-{
-public:
-    Distribution() {
-        f.resize(size);
-    }
-
-    double operator[](size_t n){
-        return this->f[n];
-    }
-
-    double operator()(double e){
-        return this->f[i_from_e(e)];
-    }
-
-    // vector-space algebra
-    Distribution& operator+=(const Distribution& d){
-        for (size_t i=0; i<size; i++) {
-            f[i] += d.f[i];
-        }
-        return *this;
-    }
-
-    Distribution& operator*=(double x){
-        for (auto& fi : f){
-            fi *= x;
-        }
-        return *this;
-    }
-
-    Distribution& operator=(const Distribution& d){
-        f = d.f;
-        return *this;
-    }
-
-    Distribution& operator=(double y){
-        for (auto& fi : f){
-            fi=y;
-        }
-        return *this;
-    }
-
-    template<typename T>
-    T expect(T (& g) (double e)){
-        T tmp = 0;
-        for (size_t i = 0; i < size; i++) {
-            tmp += g(grid[i])*f[i]*widths[i];
-        }
-        return tmp;
-    }
-
-    double eii_int (const CustomDataType::EIIdata& eii, const int i) const;
-    double tbr_int (const CustomDataType::EIIdata& eii, const int i) const;
-    void add_Qee(const Distribution& F);
-
-    // N is the (dimensionless) count of the integrated energy desired
-    void addDeltaLike(double e, double N);
-
-    // Sets the object to have a MB distribution
-    void set_maxwellian(double N, double T);
-    // defines the f interpolation
-    static void set_elec_points(size_t n, double min_e, double max_e);
-    static std::string get_energies(std::string units="eV");
-    vector<double> f;
-    static size_t size;
-private:
-    static vector<double> grid;
-
-    static vector<double> widths;
-    static double e_from_i(size_t i);
-    static size_t i_from_e(double e);
-    static double min_e, max_e;
-};
+typedef std::vector<double> bound_t; // Probabilities of state
 
 // Class responsible for storing the system state.
 class state_type
 {
 public:
-    typedef std::vector<double> bound_t; // Probabilities of state
 
     std::vector<bound_t> atomP; // Probabilities of state for all atoms.
     Distribution F; // Energy distribution function
